Banálaâáâast

+++ app/Controllers/CtrlKhachHang.php
@@ -25,12 +25,27 @@ class CtrlKhachHang extends Controller
         $nhanVien = NhanVien::where('MaNhanVien', $khachHang['MaNhanVien'])->first();
         return $this->view('Pages.KhachHangChiTiet', ['donHangS'=>$donHangS, 'khachHang'=>$khachHang, 'nhanVien'=>$nhanVien]);
     }
-    public function layDanhSachKhachHang(){
-        return KhachHang::all();
+    public function layDanhSachKhachHangTheoTrangThai($trangThai){
+        return KhachHang::where('TrangThai', $trangThai)->orderBy('MaKhachHang', 'desc')->get();
     }
     public function themKhachHang($khachHang)
-    {
+    {   
+        $lastOrder = KhachHang::orderBy('MaKhachHang', 'desc')->first();
+
+        // Nếu không có đơn hàng nào, bắt đầu với DH00000001
+        if (!$lastOrder) {
+            $newMaKhachHang = 'KH00000001';
+            $khachHang['MaKhachHang'] = $newMaKhachHang;
+        } else {
+            // Lấy mã đơn hàng cuối cùng và tăng thêm 1
+            $lastMaKhachHang = $lastOrder->MaKhachHang;
+            $numericPart = intval(substr($lastMaKhachHang, 2)); // Lấy phần số của mã
+            $newNumericPart = str_pad($numericPart + 1, 8, '0', STR_PAD_LEFT); // Tăng thêm 1 và bổ sung số 0 vào đầu
+            $newMaKhachHang = 'KH' . $newNumericPart; // Ghép lại thành mã mới
+            $khachHang['MaKhachHang'] = $newMaKhachHang;
+        }
         $result = KhachHang::insert([
+            'MaKhachHang' => $khachHang['MaKhachHang'],
             'TenKhachHang' => $khachHang['TenKhachHang'],
             'SoDienThoai' => $khachHang['SoDienThoai'],
             'GioiTinh' => $khachHang['GioiTinh'],
@@ -45,13 +60,20 @@ class CtrlKhachHang extends Controller
     }
     public function xoaKhachHang($maKhachHang)
     {
-        $result = KhachHang::where('MaKhachHang', $maKhachHang)->delete();
-        if($result){
-            return array('status' => 'success');
+        $kiemTraKhachHang = DonHang::where('MaKhachHang', $maKhachHang)->first();
+        if($kiemTraKhachHang == null){
+            $result = KhachHang::where('MaKhachHang', $maKhachHang)->delete();
+            if($result){
+                return array('status' => 'success');
+            }
+            else{
+                return array('status' => 'fail', 'message' => 'Xóa khách hàng không thành công');
+            }
         }
         else{
-            return array('status' => 'fail');
+            return array('status' => 'fail', 'message' => 'Khách hàng đã có đơn hàng không thể xóa');
         }
+
     }
     public function capNhatKhachHang($khachHang)
     {
@@ -60,6 +82,7 @@ class CtrlKhachHang extends Controller
             'SoDienThoai' => $khachHang['SoDienThoai'],
             'GioiTinh' => $khachHang['GioiTinh'],
             'DiaChi' => $khachHang['DiaChi'],
+            'TrangThai' => $khachHang['TrangThai']
         ]);
         if($result){
             return array('status' => 'success');
